// ayurvena-bootcamp/Day3-mini-ayurvena/src/types/enums.ts
export enum Gender {
    Male = "Male",
    Female = "Female",
    Other = "Other"
}
export enum Department {
    General = "General",
    Ayurveda = "Ayurveda",
    Skin = "Skin",
    Child = "Child",
    Ortho = "Ortho"
}
export enum AppointmentStatus {
    Booked = "Booked",
    Completed = "Completed",
    Cancelled = "Cancelled"
}
export enum PaymentStatus {
    Pending = "Pending",
    Paid = "Paid"
}